+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   - Added an active value that allows enabling or disabling MouseHover through the configuration.
 - Added Shared Component for pagination.
 - UserManual: Added translation for user manual which is used in Gebeudewermeplannung Portal
+- SearchBar: Render images in search result if `imagePath` is provided in SearchBarSuggestionList and SearchBarResultListGeneralItem components.
 
 ### Changed
 - Orientation: Nearby: If in an iFrame and geolocation is not enabled through config parameter, geolocation option will not show.